refactor(api): find → get にリネーム
Google AIP-131 / GitHub / Stripe など主要 API の単一リソース取得の
動詞慣習に合わせる。既存の listMedicines / createMedicine と同じ
公開 API レイヤーの語彙で揃える意図。

- findMedicineById → getMedicineById
- FindMedicineParamSchema → GetMedicineParamSchema
- findMedicineRoute → getMedicineRoute
- bruno: Find medicine → Get medicine
