+++ apps/api/bruno/medicines/Get medicine.bru
@@ -1,5 +1,5 @@
 meta {
-  name: Find medicine
+  name: Get medicine
   type: http
   seq: 3
 }
