+++ apps/api/src/schemas/medicines/index.ts
@@ -1,4 +1,4 @@
 export { CreateMedicineSchema, type CreateMedicineInput } from "./create.js";
-export { FindMedicineParamSchema, type FindMedicineParam } from "./find.js";
+export { GetMedicineParamSchema, type GetMedicineParam } from "./get.js";
 export { ListMedicinesQuerySchema, type ListMedicinesQuery } from "./list.js";
 export { TimingSchema, type Timing } from "./timing.js";
